+++ lib/searchkick/indexer.rb
@@ -24,12 +24,20 @@ module Searchkick
         # note: delete does not set error when item not found
         first_with_error = response["items"].map do |item|
           (item["index"] || item["delete"] || item["update"])
-        end.find { |item| item["error"] }
-        raise ImportError, "#{first_with_error["error"]} on item with id '#{first_with_error["_id"]}'"
+        end.find.with_index { |item, i| item["error"] && !allow_missing?(items[i], item["error"]) }
+        if first_with_error
+          raise ImportError, "#{first_with_error["error"]} on item with id '#{first_with_error["_id"]}'"
+        end
       end
 
       # maybe return response in future
       nil
+    end
+
+    private
+
+    def allow_missing?(item, error)
+      error["type"] == "document_missing_exception" && item.instance_variable_defined?(:@allow_missing)
     end
   end
 end