+++ lib/searchkick/record_indexer.rb
@@ -6,7 +6,9 @@ module Searchkick
       @index = index
     end
 
-    def reindex(records, mode:, method_name:, full: false, single: false)
+    def reindex(records, mode:, method_name:, allow_missing: false, full: false, single: false)
+      Utils.check_reindex_options(mode: mode, method_name: method_name, allow_missing: allow_missing)
+
       # prevents exists? check if records is a relation
       records = records.to_a
       return if records.empty?
@@ -51,7 +53,7 @@ module Searchkick
         index.reindex_queue.push_records(records)
       when true, :inline
         index_records, other_records = records.partition { |r| index_record?(r) }
-        import_inline(index_records, !full ? other_records : [], method_name: method_name, single: single)
+        import_inline(index_records, !full ? other_records : [], method_name: method_name, allow_missing: allow_missing, single: single)
       else
         raise ArgumentError, "Invalid value for mode"
       end
@@ -60,7 +62,7 @@ module Searchkick
       true
     end
 
-    def reindex_items(klass, items, method_name:, single: false)
+    def reindex_items(klass, items, method_name:, allow_missing: false, single: false)
       routing = items.to_h { |r| [r[:id], r[:routing]] }
       record_ids = routing.keys
 
@@ -76,7 +78,7 @@ module Searchkick
           construct_record(klass, id, routing[id])
         end
 
-      import_inline(records, delete_records, method_name: method_name, single: single)
+      import_inline(records, delete_records, method_name: method_name, allow_missing: allow_missing, single: single)
     end
 
     private
@@ -86,13 +88,13 @@ module Searchkick
     end
 
     # import in single request with retries
-    def import_inline(index_records, delete_records, method_name:, single:)
+    def import_inline(index_records, delete_records, method_name:, allow_missing:, single:)
       return if index_records.empty? && delete_records.empty?
 
       maybe_bulk(index_records, delete_records, method_name, single) do
         if index_records.any?
           if method_name
-            index.bulk_update(index_records, method_name)
+            index.bulk_update(index_records, method_name, allow_missing: allow_missing)
           else
             index.bulk_index(index_records)
           end