+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 5.2.0 (unreleased)
 
+- Added `allow_missing` option to `reindex` on relation
 - Added model name to warning about missing records
 - Fixed unnecessary data loading when reindexing relations with `:async` and `:queue` modes